refactor(demo): moves getEntity to a separate file
